requirements.txt
aes-pkcs5==1.0.2
amqp==5.2.0
asgiref==3.5.2
async-timeout==4.0.3
attrs==22.1.0
billiard==4.2.0
black==22.12.0
boto3==1.26.126
botocore==1.29.126
CacheControl==0.12.11
cachetools==5.3.0
celery==5.3.6
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.0.5
cron-descriptor==1.4.0
cryptography==41.0.5
Django==4.1.4
django-cors-headers==3.13.0
django-ebhealthcheck==2.0.2
django-filter==22.1
django-phonenumber-field==7.0.1
django-phonenumbers==1.0.1
django-rest-knox==4.2.0
django-ses==3.4.1
django-storages==1.13.2
django-timezone-field==6.1.0
djangorestframework==3.14.0
exceptiongroup==1.0.4
Faker==20.1.0
fcm-django==1.0.12
firebase-admin==5.4.0
flake8==6.0.0
google-api-core==2.11.0
google-api-python-client==2.74.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.9.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
googlemaps==4.7.3
grpcio==1.51.1
grpcio-status==1.51.1
gunicorn==20.1.0
hashids==1.3.1
haversine==2.7.0
httplib2==0.21.0
idna==3.4
iniconfig==1.1.1
jmespath==1.0.1
kombu==5.3.4
mccabe==0.7.0
msgpack==1.0.4
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==22.0
pathspec==0.10.3
phonenumbers==8.13.2
Pillow==9.4.0
platformdirs==2.6.0
pluggy==1.0.0
prompt-toolkit==3.0.41
proto-plus==1.22.2
protobuf==4.21.12
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.19.0
pyflakes==3.0.1
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
redis==5.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.3
stripe==5.0.0
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.13
vine==5.1.0
wcwidth==0.2.12
